requirements.txt
openai>=1.30.0
pydub>=0.25.1
python-dotenv>=1.0.0
azure-cognitiveservices-speech>=1.40.0
